\name{GreenCarpetProjectFiles-package}
\alias{GreenCarpetProjectFiles-package}
\alias{GreenCarpetProjectFiles}
\docType{package}
\title{
\packageTitle{GreenCarpetProjectFiles}
}
\description{
\packageDescription{GreenCarpetProjectFiles}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GreenCarpetProjectFiles}
\packageIndices{GreenCarpetProjectFiles}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GreenCarpetProjectFiles}

Maintainer: \packageMaintainer{GreenCarpetProjectFiles}
}
\keyword{ package }
